refactor: 작전시간 상태를 NormalTimer컴포너트에서 useNormalTimer로 변경
