+++ src/page/TimerPage/hooks/useNormalTimer.ts
@@ -14,6 +14,10 @@ export function useNormalTimer() {
   // 타이머가 현재 동작중인지 여부
   const [isRunning, setIsRunning] = useState(false);
 
+  // 작전 시간 타이머 작동 여부
+  const [isAdditionalTimerOn, setIsAdditionalTimerOn] = useState(false);
+  const [savedTimer, setSavedTimer] = useState(0);
+
   /**
    * 타이머를 1초마다 1씩 감소시키며 시작
    * 이미 동작중이면 재시작하지 않음
@@ -65,18 +69,57 @@ export function useNormalTimer() {
     intervalRef.current = null;
   }, [pauseTimer]);
 
+  /**
+   * 작전시간 타이머 사용 시, 기존 타이머 저장
+   */
+  const handleChangeAdditionalTimer = useCallback(() => {
+    pauseTimer();
+    if (!isAdditionalTimerOn) {
+      setSavedTimer(timer ?? 0);
+      setTimer(0);
+    } else {
+      setTimer(savedTimer);
+    }
+    setIsAdditionalTimerOn(!isAdditionalTimerOn);
+  }, [isAdditionalTimerOn, pauseTimer, setTimer, savedTimer, timer]);
+
+  /**
+   * 작전시간 타이머 사용 시, 기존 타이머 저장
+   */
+  const handleCloseAdditionalTimer = useCallback(() => {
+    setIsAdditionalTimerOn(false);
+  }, []);
+  //작전시간 종료 시, 자동으로 타이머 변경
+  useEffect(() => {
+    if (isAdditionalTimerOn && timer === 0 && isRunning) {
+      pauseTimer();
+      setTimer(savedTimer);
+      setIsAdditionalTimerOn(!isAdditionalTimerOn);
+    }
+  }, [
+    isAdditionalTimerOn,
+    timer,
+    savedTimer,
+    pauseTimer,
+    setIsAdditionalTimerOn,
+    setTimer,
+    isRunning,
+  ]);
   useEffect(() => () => pauseTimer(), [pauseTimer]);
 
   return {
     timer,
     isRunning,
     defaultTimer,
+    isAdditionalTimerOn,
     setTimer,
     startTimer,
     pauseTimer,
     resetTimer,
     setDefaultTimer,
     clearTimer,
+    handleChangeAdditionalTimer,
+    handleCloseAdditionalTimer,
   };
 }
